slight refactor of write ds; ownership is not required

+++ R/files.R
@@ -32,7 +32,7 @@ check_files_fatal <- function(x) {
 
 file_ds <- function(id = NULL) {
   ext <- ".parquet"
-  if(is.atomic(id) && !is.null(id)) {
+  if(is.atomic(id)) {
     id <- as.character(id)
     file <- paste0(.global$file.prefix, id, ext)
   } else {
@@ -44,7 +44,7 @@ file_ds <- function(id = NULL) {
 #' Move, rename, or write out data set files 
 #' 
 #' Use `move_ds()` to change the enclosing directory. `write_ds()` can also
-#' move the files, but also condenses all simulation output in to a single 
+#' move the files, but it always condenses th simulation output into a single 
 #' parquet file if multiple files are backing the mrgsimsds object. All 
 #' operations are made on the object in place; see **Details**. 
 #' 
@@ -60,10 +60,10 @@ file_ds <- function(id = NULL) {
 #' @details
 #' There is an important distinction between `write_ds()` and `move_ds()` or 
 #' `rename_ds()` for multi-file objects. The backing files can be moved or 
-#' written easily, without much computational burden. For multi-file simulation
+#' written easily, without much computational effort. For multi-file simulation
 #' outputs, `write_ds()` will need to read each file and then write the data 
 #' out to a single file. Apache Arrow can do this very efficiently, but there 
-#' will still be an additional, potentially noticeable computational burden.  
+#' will still be an additional, potentially noticeable computational effort  
 #' 
 #' When dataset files are rewritten to a single file with `write_ds()`, those 
 #' files will no longer be cleaned up when the containing R object is finalized 
@@ -72,6 +72,9 @@ file_ds <- function(id = NULL) {
 #' file cleanup will continue to occur as long as the files remain under 
 #' `tempdir()`. No change in finalization behavior due to garbage collection 
 #' of the containing object will happen when files are renamed. 
+#' 
+#' The object (`x`) is required to own the underlying files in order to move
+#' them with `move_ds()`; ownership is not required for `write_ds()`.
 #' 
 #' All three functions modify `x` in place and file ownership stays with `x`. 
 #' 
@@ -143,15 +146,18 @@ rename_ds <- function(x, id) {
 #' @export
 write_ds <- function(x, sink, ...) {
   require_ds(x)
-  if(!check_ownership(x)) {
-    abort("cannot re-write files you don't own.")  
+  owner <- check_ownership(x)
+  if(owner) {
+    disown(x)
+    mv_or_cp <- file_move  
+  } else {
+    mv_or_cp <- file_copy
   }
-  disown(x)
   if(length(x$files)==1) {
-    file_move(x$files, sink)
+    mv_or_cp(x$files, sink)
   } else {
     write_parquet(x$ds, sink, ...)
-    unlink(x$ds$files, recursive = TRUE)
+    if(owner) unlink(x$ds$files, recursive = TRUE)
   }
   x$gc <- FALSE
   x$files <- sink
